Switch to bigint field

+++ priv/repo/migrations/20180824204501_create_uploads_table.exs
@@ -5,7 +5,7 @@ defmodule Ret.Repo.Migrations.CreateUploadsTable do
     create table(:uploads, prefix: "ret0", primary_key: false) do
       add(:upload_id, :bigint, default: fragment("ret0.next_id()"), primary_key: true)
       add(:upload_uuid, :string)
-      add(:uploader_account_id, :integer, null: false)
+      add(:uploader_account_id, :bigint, null: false)
       # TODO BP: These should probably be non-null
       add(:state, :string)
       add(:size, :bigint)
